+++ packages/dropdown/src/index.test.tsx
@@ -1,12 +1,12 @@
 import * as React from 'react';
 import { mount } from 'enzyme';
 import Theme from '@lendi-ui/theme';
-import InlineDropdown from './index';
+import Dropdown from './index';
 import { DropdownWrapper, Select, IconDown } from './index.style';
 
 let Wrapper;
 
-describe('InlineDropdown', () => {
+describe('Dropdown', () => {
   beforeEach(() => {
     const items = [
       {
@@ -33,22 +33,22 @@ describe('InlineDropdown', () => {
 
     Wrapper = mount(
       <Theme>
-        <InlineDropdown size="md" items={items} />
+        <Dropdown size="md" items={items} />
       </Theme>
     );
   });
 
   it('render whole component', () => {
-    expect(Wrapper.find('InlineDropdown').length).toEqual(1);
-    expect(Wrapper.find('InlineDropdown')).toMatchSnapshot();
+    expect(Wrapper.find('Dropdown').length).toEqual(1);
+    expect(Wrapper.find('Dropdown')).toMatchSnapshot();
   });
 
-  it('render SelectedBox component and has size property "md"', () => {
+  it('render DropdownWrapper', () => {
     expect(Wrapper.find(DropdownWrapper).length).toEqual(1);
     expect(Wrapper.find(DropdownWrapper)).toMatchSnapshot();
   });
 
-  it('render SelectedBox component and has size property "md"', () => {
+  it('render Dropdown component and has size property "md"', () => {
     expect(Wrapper.find(Select).length).toEqual(1);
     expect(Wrapper.find(Select).prop('selectSize')).toEqual('md');
     expect(Wrapper.find(Select)).toMatchSnapshot();