+++ packages/inline-dropdown/CHANGELOG.md
@@ -1,1 +1,5 @@
-1.1.0
+# Changelog - @lendi-ui/inline-dropdown
+
+## [1.1.0] - 2019-02-11 - [BRAVO-456](https://creditandfinance.atlassian.net/browse/BRAVO-456)
+### Deprecated
+- component deprecated in favor of a more generic @lendi-ui/dropdown